backend/requirements.txt
django-cors-headers
Django
django-allauth
djangorestframework
psycopg2

